refactor: complete the dedup sweep in telemetry integration tests
Addresses review findings from the reviewer battery (lazy-checker,
wtf-reviewer, integration-reviewer, nitpicker all converged on the same
gaps):

- short_timeout_db fixture now uses open_db_with_session() instead of
  hand-rolling the DB-init/session-insert sequence it was extracted to
  eliminate
- open_db_with_session()'s return annotation uses the DbFixture alias
  it's paired with everywhere else
- assert_last_error_row_coherence() uses only_row() instead of a bare
  results[0], matching every other single-row lookup in the directory
- Two leftover rows[0] lookups in test_job_queries.py switched to
  only_row(), restoring the cardinality assertion the rest of the file
  already has
- _fetch_blocking_events/_drain and _fetch_blocking_events/_drain_tasks
  (byte-identical private helpers duplicated across
  test_blocking_events_persistence.py and
  test_blocking_io_executor_offload.py) consolidated into
  helpers.fetch_blocking_events/drain_db_writes
- dup-ignore annotations added to the three near-identical
  register-a-live-job test bodies in
  test_global_jobs_and_service_info.py, per the directory's documented
  convention for genuinely parallel test structure
